(WIP #34) Refactoring Modules
